# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This API made for Jeres portfolio",
        "version": "2.0.0",
        "title": "Jeres API",
        "contact": {
            "email": "canlacelis@gmail.com",
            "mobile": "09959003488"
        }
    },
    "schemes": [
        "https"
    ],
    "basePath": "https://jeres-api-pb.herokuapp.com/",
    "paths": {
        "/skills": {
            "get": {
                "summary": "get all skills information",
                "description": "get all skill name, skill id, and skill description. ",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Successful get",
                        "schema": {
                            "type": "array"
                        }
                    },
                    "400": {
                        "description": "Failed"
                    }
                }
            },
            "post": {
                "summary": "add one skill information",
                "description": "add skill name, skill id, and skill description. ",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Successful get",
                        "schema": {
                            "type": "array"
                        }
                    },
                    "400": {
                        "description": "Failed"
                    }
                }
            }
        },
        "/skills/{id}": {
            "delete": {
                "summary": "delete one skill using id",
                "description": "hard delete the skill. ",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted"
                    }
                }
            }
        }
    }
}